requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29

langchain>=0.2
langchain-core>=0.2
langchain-community>=0.2
langchain-groq>=0.1

langgraph>=0.1

langchain-huggingface>=0.0.3
sentence-transformers>=2.6

langchain-chroma>=0.1
chromadb>=0.4

tavily-python>=0.3

python-dotenv>=1.0
pydantic>=2.0
requests>=2.31

streamlit>=1.30

torch>=2.0